Faithful title menu: reconstruct MenuComponent/SelectableGroup/Selectable, retire the facade
Replaces the static-array MenuComponent facade (BrnBootLegalBoundary) with the real
X360 frontend classes so BootLegal drives the title menu exactly as the console does:

- BrnGui::Selectable: the 4 state-flag setters (SetActive/SetHighlightable/
  SetSelectable/SetHighlighted, change-detect + dirty), Update, Construct.
- BrnGui::SelectableGroup: full method family (Construct/Clear/Add/GetSelectable/
  GetHighlighted/HighlightIndex/HighlightNext/HighlightPrevious/Enable/Disable/
  Select/Update/SetIds/SetSelectableId/GetIndexFromId/HighlightId/GetHighlightedId).
  maSelectables retyped u32 -> Selectable* (x64 pointer-correct); GuiComponent
  identity modelled by composition.
- BrnGui::MenuComponent: real object model (SelectableGroup + MenuItem maItems[16] +
  miNumMenuItems) + ctor, Construct, SetupMenu, SetText, Clear, GetSelectable, SetId,
  DisableSelectable, HighlightNext/Previous, AppendExpectedAptComponent.
- BootLegal wired to the real methods (Refresh -> Update/Clear, SelectNext/Previous
  -> HighlightNext/Previous, highlighted index read off the group) per the X360
  menu-vtable slots (verified from asm: +0x14 Update, +0x18 Clear, +0x34/+0x38 nav).

Item apt output still flows through GuiComponent::AddOutputAptViewState (the deeper
AptCommunicator seam remains the next milestone). All affected TUs compile-gate clean.

+++ src/GameSource/Gui/Flow/Shared/Components/BrnMenuToggleGroup.cpp
@@ -177,8 +177,7 @@ namespace BrnGui
     {
         CGS_ASSERT(liIndex >= 0 && liIndex < TI_SIZE,
                    "MenuToggleGroup::GetSelectable() invalid index specified");
-        return reinterpret_cast<MenuToggle*>(
-            static_cast<uintptr_t>(SelectableGroup::GetSelectable(liIndex)));
+        return reinterpret_cast<MenuToggle*>(SelectableGroup::GetSelectable(liIndex));
     }
 
     // ---------------------------------------------------------------------------------
